refactor: collapse metering/cost hooks into unified Metering.install_hook
Delete hooks/metering.rb and hooks/cost_tracking.rb. Add Metering.install_hook
that does a single extraction then records in Metering::Recorder AND emits via
Metering.emit (llm.metering exchange). The old lex.metering.record routing key
was a dead gateway artifact and is intentionally dropped. Update hooks.rb to
call Metering.install_hook from install_defaults.
